added outputs, added kill application, renamed file
